+++ src/host_bsp.c
@@ -26,17 +26,19 @@ see the files COPYING and COPYING.LESSER. If not, see
 #include "common.h"
 
 #include <stdio.h>
-#include <unistd.h>
 #include <string.h>
 #include <stdlib.h>
-#include <time.h>
 #include <stddef.h>
+// We need to do this in order to use the timers that give wall time
+#define __USE_POSIX199309
+#include <time.h>
 
 // Global state
 bsp_state_t state;
 int bsp_initialized = 0;;
 
 void _host_sync();
+void _microsleep(int microseconds); //1000 gives 1 millisecond
 void _get_p_coords(int pid, int* row, int* col);
 bsp_state_t* _get_state();
 
@@ -133,7 +135,7 @@ int bsp_begin(int nprocs)
     // Then the functions that DID succeed should be undone again
     // So at e_load_group failure it cleanup the e_open result
 
-    if (nprocs < 1 || nprocs > 99 || nprocs > MAX_NCORES) {
+    if (nprocs < 1 || nprocs > _NPROCS) {
         fprintf(stderr, "ERROR: nprocs = %d.\n", nprocs);
         return 0;
     }
@@ -219,11 +221,11 @@ int ebsp_spmd()
         return 0;
     }
 
-    int i, j;
+    int i;
     int cores_initialized;
     while (1)
     {
-        usleep(1000); //0.01 seconds
+        _microsleep(1000); // 1 millisecond
 
         // Read the full communication buffer
         if (e_read(&state.emem, 0, 0, 0, &state.comm_buf,
@@ -245,26 +247,31 @@ int ebsp_spmd()
     printf("(BSP) DEBUG: All epiphany cores are ready for initialization.\n");
 #endif
 
-    //All cores are waiting.
-    //We can now send data and 'start' them
-    clock_t start = clock(); 
-    clock_t end = clock(); 
-    float arm_timer;
-    arm_timer = (float)(end-start) / ARM_CLOCKSPEED;
+    // Time storage
+    struct timespec ts_start, ts_end;
+    float time_elapsed;
+
+    // Starting time
+    clock_gettime(CLOCK_MONOTONIC, &ts_start);
+    // Current time. Repeat these two lines every iteration
+    clock_gettime(CLOCK_MONOTONIC, &ts_end);
+    time_elapsed = (ts_end.tv_sec - ts_start.tv_sec + (ts_end.tv_nsec - ts_start.tv_nsec) * 1.0e-9);
+
+    // All cores are waiting.
+    // We can now send data and 'start' them
+    _write_extmem(&time_elapsed,
+            offsetof(ebsp_comm_buf, remotetimer),
+            sizeof(float));
     for (i = 0; i < state.nprocs; ++i)
     {
-        //Core i has written its coredata address
-        //Now we can initialize it
-        _write_coredata(i,
-                &arm_timer,
-                offsetof(ebsp_core_data, remotetimer),
-                sizeof(int));
+        // Core i has written its coredata address
+        // Now we can initialize it
         _write_coredata(i,
                 &state.nprocs,
                 offsetof(ebsp_core_data, nprocs),
                 sizeof(int));
 
-        //All data is initialized. Send start signal
+        // All data is initialized. Send start signal
         int flag = STATE_CONTINUE;
         _write_coredata(i,
                 &flag,
@@ -273,6 +280,7 @@ int ebsp_spmd()
     }
 
     int total_syncs = 0;
+    int extmem_corrupted = 0;
     int run_counter;
     int sync_counter;
     int finish_counter;
@@ -284,14 +292,14 @@ int ebsp_spmd()
 #endif
 
     while (finish_counter != state.nprocs) {
-        //Write timer
-        end=clock(); 
-        arm_timer = (float)(end-start)/ARM_CLOCKSPEED;
-        for(i = 0; i < state.nprocs; i++)
-            _write_coredata(i,
-                    &arm_timer,
-                    offsetof(ebsp_core_data, remotetimer),
-                    sizeof(int));
+        // Update timer
+        clock_gettime(CLOCK_MONOTONIC, &ts_end);
+        time_elapsed = (ts_end.tv_sec - ts_start.tv_sec + (ts_end.tv_nsec - ts_start.tv_nsec) * 1.0e-9);
+        _write_extmem(&time_elapsed,
+                offsetof(ebsp_comm_buf, remotetimer),
+                sizeof(float));
+
+        _microsleep(100); //1000 is 1 millisecond
 
         // Read the full communication buffer
         // Including pushreg states and so on
@@ -302,7 +310,7 @@ int ebsp_spmd()
             return 0;
         }
 
-        //Check sync states
+        // Check sync states
         run_counter      = 0;
         sync_counter     = 0;
         finish_counter   = 0;
@@ -313,21 +321,23 @@ int ebsp_spmd()
                 case STATE_SYNC:     sync_counter++; break;
                 case STATE_FINISH:   finish_counter++; break;
                 case STATE_CONTINUE: continue_counter++; break;
-                default: fprintf(stderr, "ERROR: syncstate[%d] = %d.\n",
-                                 i, state.comm_buf.syncstate[i]);
+                default: extmem_corrupted++;
+                         if( extmem_corrupted <= 32) //to avoid output overflow
+                             fprintf(stderr, "ERROR: External memory corrupted. syncstate[%d] = %d.\n",
+                                     i, state.comm_buf.syncstate[i]);
                          break;
             }
             if (state.comm_buf.msgflag[i])
             {
                 printf("$%02d: %s\n", i, state.comm_buf.message[i].msg);
-                //Reset flag so we do not read it again
+                // Reset flag so we do not read it again
                 state.comm_buf.msgflag[i] = 0;
-                //Write the int to the external comm_buf
+                // Write the int to the external comm_buf
                 _write_extmem(&state.comm_buf.msgflag[i],
                         offsetof(ebsp_comm_buf, msgflag[i]),
                         sizeof(int));
-                //Signal epiphany core that message was read
-                //so that it can (possibly) output the next message
+                // Signal epiphany core that message was read
+                // so that it can (possibly) output the next message
                 _write_coredata(i, &state.comm_buf.msgflag[i],
                         offsetof(ebsp_core_data, msgflag), sizeof(int));
             }
@@ -346,9 +356,9 @@ int ebsp_spmd()
         if (sync_counter == state.nprocs) {
             ++total_syncs;
 #ifdef DEBUG
-            //This part of the sync (host side)
-            //usually does not crash so only one
-            //line of debug output is needed here
+            // This part of the sync (host side)
+            // usually does not crash so only one
+            // line of debug output is needed here
             printf("(BSP) DEBUG: Sync %d\n", total_syncs);
 #endif
             _host_sync();
@@ -357,23 +367,22 @@ int ebsp_spmd()
             if (state.sync_callback)
                 state.sync_callback();
 
-            //First reset the comm_buf
+            // First reset the comm_buf
             for (i = 0; i < state.nprocs; i++)
                 state.comm_buf.syncstate[i] = STATE_CONTINUE;
             _write_extmem(&state.comm_buf.syncstate,
                     offsetof(ebsp_comm_buf, syncstate),
                     _NPROCS * sizeof(int));
-            //Now write it to all cores to continue their execution
+            // Now write it to all cores to continue their execution
             for (i = 0; i < state.nprocs; i++)
                 _write_coredata(i, &state.comm_buf.syncstate[i],
                         offsetof(ebsp_core_data, syncstate), sizeof(int));
         }
-
-        usleep(1000);
     }
     printf("(BSP) INFO: Program finished\n");
 
-    state.end_callback();
+    if (state.end_callback)
+        state.end_callback();
 
     return 1;
 }
@@ -414,11 +423,11 @@ int bsp_nprocs()
 // Memory
 void _host_sync() {
     // TODO: Right now bsp_pop_reg is ignored
-    int i, j;
 
     // Check if core 0 did a push_reg
     if (state.comm_buf.pushregloc[0] != NULL)
     {
+        int i;
 #ifdef DEBUG
         printf("(BSP) DEBUG: bsp_push_reg occurred. core 0 addr = %p\n",
                 state.comm_buf.pushregloc[0]);
@@ -456,6 +465,16 @@ void _host_sync() {
                 offsetof(ebsp_comm_buf, pushregloc),
                 _NPROCS * sizeof(void*));
     }
+}
+
+void _microsleep(int microseconds)
+{
+    // This function itself is about 300 microseconds overhead
+    struct timespec request, remain;
+    request.tv_sec = (int)(microseconds / 1000000);
+    request.tv_nsec = (microseconds - 1000000 * request.tv_sec) * 1000;
+    if (clock_nanosleep(CLOCK_MONOTONIC, 0, &request, &remain) != 0)
+        fprintf(stderr, "ERROR: clock_nanosleep was interrupted.");
 }
 
 void _get_p_coords(int pid, int* row, int* col)