+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 0.2
+- Fixed timer to use real time instead of clockcycles
 - Added `ebsp_message`
 - External memory in one large block
 - Fixed hardcoded BSP addresses to no longer overlap